backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
pydriller==2.6
radon==6.0.1
langchain==0.2.0
langchain-openai==0.1.7
langgraph==0.1.19
openai==1.30.0
PyGithub==2.3.0
gitpython==3.1.43
jinja2==3.1.4
weasyprint==62.3
matplotlib==3.9.0
scikit-learn==1.5.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.7.1
markdown==3.6
reportlab==4.2.2

